refactor: separate config ui into own components
